move list component
